Rename StrpPolicy.cs to SrtpPolicy.cs
Name mismatch
